277 Items from AICP Spreadsheet Added and Data Types are also now in the table
When you load items, now the 277 items from the AICP spreadsheet are inserted into the table.
The data types associated with each data field is now also listed in the database and on
the table. By that I mean something like 'Days @ Location' and 'Distance Traveled' are displayed
above a data field